refactor(audio and playist/list): extract list logic to livewire component
