+++ resources/views/home.blade.php
@@ -20,31 +20,5 @@
     #audios {
         height: 70vh;
     }
-
-    .wrap {
-        flex-wrap: wrap;
-    }
-
-    .audio-card {
-        width: 20%;
-        min-width: 20%;
-    }
-
-    .playlist-card {
-        width: 20%;
-        min-width: 20%;
-    }
-
-    @media screen and (max-width: 800px) {
-        .audio-card {
-            width: 100%;
-            min-width: 100%;
-        }
-
-        .playlist-card {
-            width: 100%;
-            min-width: 100%;
-        }
-    }
 </style>
 @endpush
